move under `optimize` namespace
