+++ pyiceberg/table/inspect.py
@@ -16,12 +16,11 @@
 # under the License.
 from __future__ import annotations
 
-from datetime import datetime, timedelta, timezone
+from datetime import datetime, timezone
 from functools import reduce
 from typing import TYPE_CHECKING, Any, Dict, Iterator, List, Optional, Set, Tuple, Union, cast
 
 from pyiceberg.conversions import from_bytes
-from pyiceberg.io import _parse_location
 from pyiceberg.manifest import DataFile, DataFileContent, ManifestContent, PartitionFieldSummary
 from pyiceberg.partitioning import PartitionSpec
 from pyiceberg.table.snapshots import Snapshot, ancestors_of
@@ -687,40 +686,3 @@ class InspectTable:
         _all_known_files["datafiles"] = reduce(set.union, files_by_snapshots, set())
 
         return _all_known_files
-
-    def orphaned_files(self, location: str, older_than: timedelta = timedelta(days=3)) -> Set[str]:
-        """Get all the orphaned files in the table.
-
-        Args:
-            location: The location to check for orphaned files.
-            older_than: The time period to check for orphaned files. Defaults to 3 days.
-
-        Returns:
-            A set of orphaned file paths.
-
-        """
-        try:
-            import pyarrow as pa  # noqa: F401
-        except ModuleNotFoundError as e:
-            raise ModuleNotFoundError("For deleting orphaned files PyArrow needs to be installed") from e
-
-        from pyarrow.fs import FileSelector, FileType
-
-        from pyiceberg.io.pyarrow import _fs_from_file_path
-
-        all_known_files = self.all_known_files()
-        flat_known_files: set[str] = reduce(set.union, all_known_files.values(), set())
-
-        fs = _fs_from_file_path(self.tbl.io, location)
-
-        _, _, path = _parse_location(location)
-        selector = FileSelector(path, recursive=True)
-        # filter to just files as it may return directories, and filter on time
-        as_of = datetime.now(timezone.utc) - older_than
-        all_files = [
-            f.path for f in fs.get_file_info(selector) if f.type == FileType.File and (as_of is None or (f.mtime < as_of))
-        ]
-
-        orphaned_files = set(all_files).difference(flat_known_files)
-
-        return orphaned_files
